% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail_function.R
\name{rail_function}
\alias{rail_function}
\title{A Rail Function}
\usage{
rail_function(love = TRUE)
}
\arguments{
\item{love}{Do you love rails? Defaults to TRUE.}
}
\description{
This function allows you to express your love of rails.
}
\examples{
rail_function()





}
\keyword{rails}

